+++ vampytest/__main__.py
@@ -1,4 +1,12 @@
-import sys
+from os import getcwd as get_current_working_directory
+from sys import argv as COMMAND_LINE_PARAMETERS, path as SYSTEM_PATHS
+
+# When you run `python3 -m vampytest`, the current working directory is added as the first parameter;
+# We do not want that, because it makes local directories importable before other packages.
+CURRENT_WORKING_DIRECTORY = get_current_working_directory()
+if (len(SYSTEM_PATHS) > 1) and (SYSTEM_PATHS[0] == CURRENT_WORKING_DIRECTORY):
+    del SYSTEM_PATHS[0]
+
 
 from scarletio import get_short_executable
 
@@ -9,7 +17,7 @@ def __main__():
     """
     Executes vampytest from terminal.
     """
-    parameters = sys.argv.copy()
+    parameters = COMMAND_LINE_PARAMETERS.copy()
     
     if parameters and ((parameters[0] == __file__) or (parameters[0] != get_short_executable())):
         del parameters[0]