+++ vampytest/core/environment/default.py
@@ -29,7 +29,7 @@ class DefaultEnvironment(RichAttributeErrorBaseType):
     
     def __repr__(self):
         """Returns the environment's representation."""
-        return f'<{self.__class__.__name__}>'
+        return f'<{type(self).__name__}>'
     
     
     def run(self, test, positional_parameters, keyword_parameters):