+++ vampytest/core/handling/handle.py
@@ -75,7 +75,7 @@ class Handle(RichAttributeErrorBaseType):
     
     def __repr__(self):
         """Returns the test handle's representation."""
-        repr_parts = ['<', self.__class__.__name__]
+        repr_parts = ['<', type(self).__name__]
         
         repr_parts.append(' test = ')
         repr_parts.append(short_repr(self.test))