+++ vampytest/core/wrappers/wrapper_conflict.py
@@ -40,7 +40,7 @@ class WrapperConflict(RichAttributeErrorBaseType):
     
     def __repr__(self):
         """Returns the wrapper conflict's representation."""
-        repr_parts = ['<', self.__class__.__name__]
+        repr_parts = ['<', type(self).__name__]
         
         repr_parts.append(' wrapper_0 = ')
         repr_parts.append(repr(self.wrapper_0))